Update huawei telnet driver support
