+++ netmiko/huawei/__init__.py
@@ -1,5 +1,5 @@
 from __future__ import unicode_literals
-from netmiko.huawei.huawei_ssh import HuaweiSSH, HuaweiVrpv8SSH
-from netmiko.huawei.huawei_telnet import HuaweiTelnet
+from netmiko.huawei.huawei import HuaweiSSH, HuaweiVrpv8SSH
+from netmiko.huawei.huawei import HuaweiTelnet
 
 __all__ = ["HuaweiSSH", "HuaweiVrpv8SSH", "HuaweiTelnet"]
